update LICENSE to MIT instead of Apache
I am still developing the rationale and justification but I want to have all frontend web-browser side code be MIT licensed and prefer instead BUSL for the service and customer-facing bits.

This allows for developing in the open while having some protection against another business using all my hard work to easily eat my lunch.

It's not that Apache is incompatible so much as it is easier to manage if I'm consistent across all related repositories.